Work in progress check-in for bug #21445 "Improve Thrid Person Avatar Gaze"
The way desktop 3rd person control of the Avatar head works currently:
MyAvatar::updateOrientation() takes controler input and calculates
delta rotation for the avatar (yaw for the whole body, pitch for the head).
Application::paintGL() then takes the orientation of the avatar's head
and applies that to the camera.

To allow the avatar and camera to look at different things in this mode,
I added vars and accessors for the delta rotational angles (pitch and yaw)
calculated in MyAvatar::updateOrientation() so that Application::paintGL()
could use them to rotate the camera directly instead of being slaved to
the avatar head. (although to avoid the problem of not being able to
initialize the camera's orientation with the avatar if the game starts in
3rd Person mode, I wound up still slaving the camera's yaw to the avatar)

I then modified the 3rd person desktop case in MyAvatar::updateOrientation()
to cast a forward ray from the center of the camera and have the avatar head look
at the point of intersection (if there is one)

In practice, this produces very similar head rotation behavior to the
existing behavior, only differing in the case where the 3rd person camera
isn't centered on the avatar, but even then it's not a huge difference.